requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.3.1
certifi==2025.1.31
cffi==1.17.1
cryptography==44.0.2
distro==1.9.0
grpcio==1.71.0
grpcio-health-checking==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.2.4
openai==1.66.5
protobuf==5.29.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
rich==13.9.4
setuptools==76.1.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
validators==0.34.0
weaviate==0.1.2
weaviate-client==4.11.1
